execution/requirements.txt
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
python-dotenv
playwright
beautifulsoup4
requests
